treat clients list properly through build too
