+++ packages/react-email/src/commands/dev.ts
@@ -1,36 +1,12 @@
 import fs from 'node:fs';
 import { setupHotreloading, startDevServer } from '../utils/index.js';
+import { validateEmailClientList } from '../utils/validate-email-client-list.js';
 
 interface Args {
   dir: string;
   port: string;
   clients: string;
 }
-
-const validEmailClients = [
-  'gmail',
-  'outlook',
-  'yahoo',
-  'apple-mail',
-  'aol',
-  'thunderbird',
-  'microsoft',
-  'samsung-email',
-  'sfr',
-  'orange',
-  'protonmail',
-  'hey',
-  'mail-ru',
-  'fastmail',
-  'laposte',
-  't-online-de',
-  'free-fr',
-  'gmx',
-  'web-de',
-  'ionos-1and1',
-  'rainloop',
-  'wp-pl',
-] as const;
 
 export const dev = async ({
   dir: emailsDirRelativePath,
@@ -43,15 +19,7 @@ export const dev = async ({
       process.exit(1);
     }
 
-    const clientsArray = clients.split(/\s*,\s*/);
-    for (const client of clientsArray) {
-      if (!(validEmailClients as readonly string[]).includes(client)) {
-        console.error(
-          `${client} is not one of the clients we support. Valid options are ${validEmailClients.join(', ')}`,
-        );
-        process.exit(1);
-      }
-    }
+    const clientsArray = validateEmailClientList(clients);
 
     const devServer = await startDevServer(
       emailsDirRelativePath,
